requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12

torch
torchvision
numpy
Pillow
opencv-python-headless==4.10.0.84
grad-cam==1.5.4

pytest==8.3.3
httpx==0.27.2
